+++ cw-orch/tests/querier.rs
@@ -167,6 +167,7 @@ mod queriers {
     }
 
     #[test]
+    #[serial_test::serial]
     fn simulate_tx() {
         let rt = Runtime::new().unwrap();
 
@@ -210,26 +211,30 @@ mod queriers {
     }
 
     #[test]
+    #[serial_test::serial]
     fn contract_info() {
+        use crate::common::Id;
+        use cw_orch::prelude::networks;
+
         let rt = Runtime::new().unwrap();
         let channel = rt.block_on(build_channel());
         let cosm_wasm = CosmWasm::new(channel);
-
-        use cw_orch::prelude::networks;
-
-        let runtime = tokio::runtime::Runtime::new().unwrap();
-
         let daemon = Daemon::builder()
             .chain(networks::LOCAL_JUNO)
-            .handle(runtime.handle())
+            .handle(rt.handle())
             .build()
             .unwrap();
 
         let sender = daemon.sender();
 
-        let contract = mock_contract::MockContract::new("test:mock_contract", daemon.clone());
+        let contract = mock_contract::MockContract::new(
+            format!("test:mock_contract:{}", Id::new()),
+            daemon.clone(),
+        );
 
-        let _ = contract.instantiate(&InstantiateMsg {}, Some(&sender), None);
+        contract
+            .instantiate(&InstantiateMsg {}, Some(&sender), None)
+            .unwrap();
 
         let contract_address = contract.address().unwrap();
